# 34% LLaMa speed