175.67/131.16=1.34 -> 34% faster

Easier to review commit-by-commit.

A stack of performance improvements, mainly.
* set of `WeakRef` instead of a (slow) `WeakRefSet`
* faster handling of buffer info (not recalculating memsize)
* views and shapetrackers expose whether they are symbolic or not (now `view.is_all_int`, could be `view.symbolic` (@chenyuxyz what do you think of this?)
* caches on shapetracker

I suggest splitting this up in one MR per commit. There are some gains left by moving things from `shapetracker.py` to `view.py` that belong there, such as the `merge_view` logic, but then this diff would explode.

Timed on M2 Pro.

Before (0eb0defa6f72eed5c75c8ffe16745d7b9904341a):
```
using METAL backend
using LLaMA-7B model
ram used: 13.48 GB, freqs_cis                                         : 100%|███████████████████████████████████████████████████████████████████████████████████████████████| 292/292 [00:02<00:00, 109.22it/s]
loaded weights in 2681.46 ms, 13.48 GB loaded at 5.03 GB/s
Hello.
ran model in 286.69 ms
total 881.30 ms, 1.13 tok/sec
 I
ran model in 195.53 ms
total 198.68 ms, 5.03 tok/sec
'
ran model in 188.31 ms
total 192.09 ms, 5.21 tok/sec
m
ran model in 176.73 ms
total 180.47 ms, 5.54 tok/sec
 a
ran model in 176.38 ms
total 179.07 ms, 5.58 tok/sec
 
ran model in 175.67 ms
total 179.38 ms, 5.57 tok/sec
2
ran model in 177.33 ms
total 181.07 ms, 5.52 tok/sec
0
ran model in 199.06 ms
total 201.85 ms, 4.95 tok/sec
 year
ran model in 177.85 ms
total 181.59 ms, 5.51 tok/sec
 old
ran model in 179.50 ms
total 183.27 ms, 5.46 tok/sec
```


After:
```
using METAL backend
using LLaMA-7B model
ram used: 13.48 GB, freqs_cis                                         : 100%|████████████████████████████████████████████████████████████████| 292/292 [00:02<00:00, 132.27it/s]
loaded weights in 2215.31 ms, 13.48 GB loaded at 6.08 GB/s
Hello.
ran model in 256.14 ms
total 685.87 ms, 1.46 tok/sec
 I
ran model in 167.88 ms
total 170.09 ms, 5.88 tok/sec
'
ran model in 131.16 ms
total 134.32 ms, 7.44 tok/sec
m
ran model in 150.92 ms
total 154.02 ms, 6.49 tok/sec
 a
ran model in 150.25 ms
total 154.25 ms, 6.48 tok/sec
 
ran model in 135.83 ms
total 138.93 ms, 7.20 tok/sec
2
ran model in 152.59 ms
total 155.65 ms, 6.42 tok/sec
0
ran model in 136.45 ms
total 140.21 ms, 7.13 tok/sec
 year
ran model in 154.36 ms
total 157.38 ms, 6.35 tok/sec
 old
ran model in 131.42 ms
total 134.59 ms, 7.43 tok/sec
```