@geohot 2023-11-21
This speed currently isn't worth the complexity. Just sticking cache everywhere is not a change I'm going to merge. Is there a way to gain 20% that also makes the code cleaner?